requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.6.0
joblib>=1.4.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pyodbc>=5.1.0
python-dotenv>=1.0.1
